backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
akshare==1.11.64
yfinance==0.2.33
python-dateutil==2.8.2
Werkzeug==3.0.1
click==8.1.7
